+++ xarray/tests/test_backends.py
@@ -2170,6 +2170,68 @@ class TestRasterio(TestCase):
                     with xr.open_dataarray(tmp_nc_file) as ncds:
                         assert_identical(rioda, ncds)
 
+    @requires_scipy_or_netCDF4
+    def test_nodata(self):
+        import rasterio
+        from rasterio.transform import from_origin
+
+        # Create a geotiff file in utm proj
+        with create_tmp_file(suffix='.tif') as tmp_file:
+            # data
+            nx, ny, nz = 4, 3, 3
+            data = np.arange(nx*ny*nz,
+                             dtype=rasterio.float32).reshape(nz, ny, nx)
+            transform = from_origin(5000, 80000, 1000, 2000.)
+            with rasterio.open(
+                    tmp_file, 'w',
+                    driver='GTiff', height=ny, width=nx, count=nz,
+                    crs={'units': 'm', 'no_defs': True, 'ellps': 'WGS84',
+                         'proj': 'utm', 'zone': 18},
+                    transform=transform,
+                    nodata=-9765,
+                    dtype=rasterio.float32) as s:
+                s.write(data)
+            expected_nodatavals = [-9765, -9765, -9765]
+            with xr.open_rasterio(tmp_file) as rioda:
+                np.testing.assert_array_equal(rioda.attrs['nodatavals'],
+                                              expected_nodatavals)
+                with create_tmp_file(suffix='.nc') as tmp_nc_file:
+                    rioda.to_netcdf(tmp_nc_file)
+                    with xr.open_dataarray(tmp_nc_file) as ncds:
+                        np.testing.assert_array_equal(ncds.attrs['nodatavals'],
+                                                      expected_nodatavals)
+
+    @requires_scipy_or_netCDF4
+    def test_nodata_missing(self):
+        import rasterio
+        from rasterio.transform import from_origin
+
+        # Create a geotiff file in utm proj
+        with create_tmp_file(suffix='.tif') as tmp_file:
+            # data
+            nx, ny, nz = 4, 3, 3
+            data = np.arange(nx*ny*nz,
+                             dtype=rasterio.float32).reshape(nz, ny, nx)
+            transform = from_origin(5000, 80000, 1000, 2000.)
+            with rasterio.open(
+                    tmp_file, 'w',
+                    driver='GTiff', height=ny, width=nx, count=nz,
+                    crs={'units': 'm', 'no_defs': True, 'ellps': 'WGS84',
+                         'proj': 'utm', 'zone': 18},
+                    transform=transform,
+                    dtype=rasterio.float32) as s:
+                s.write(data)
+
+            expected_nodatavals = [np.nan, np.nan, np.nan]
+            with xr.open_rasterio(tmp_file) as rioda:
+                np.testing.assert_array_equal(rioda.attrs['nodatavals'],
+                                              expected_nodatavals)
+                with create_tmp_file(suffix='.nc') as tmp_nc_file:
+                    rioda.to_netcdf(tmp_nc_file)
+                    with xr.open_dataarray(tmp_nc_file) as ncds:
+                        np.testing.assert_array_equal(ncds.attrs['nodatavals'],
+                                                      expected_nodatavals)
+
     def test_utm(self):
         import rasterio
         from rasterio.transform import from_origin